임포트 기능 추가

+++ .example/models/user.go
@@ -10,7 +10,7 @@ type User struct {
 	Name        string    `description:"이름"`
 	Description string    `description:"설명" sql:"type:text"`
 	Owner       string    `description:"주인"`
-	CarID       uint64    `description:"CarID입니다"`
+	CarID       uint      `description:"CarID입니다"`
 }
 
 func init() {
